refactor(message-list): move LocalDeleteOperationDecider to feature module
Migrate LocalDeleteOperationDecider from legacy/core to
feature/mail/message/list as an interface (api) with
DefaultLocalDeleteOperationDecider implementation (internal).
Wire DI through FeatureMessageListModule.

Fixes #10515
