+++ legacy/core/src/test/java/com/fsck/k9/controller/MessagingControllerTest.java
@@ -39,10 +39,12 @@ import com.fsck.k9.notification.NotificationStrategy;
 import net.thunderbird.core.common.mail.Protocols;
 import net.thunderbird.core.logging.Logger;
 import net.thunderbird.core.outcome.Outcome;
+import net.thunderbird.feature.mail.message.list.LocalDeleteOperationDecider;
 import net.thunderbird.feature.mail.folder.api.OutboxFolderManager;
 import net.thunderbird.feature.notification.api.NotificationManager;
 import net.thunderbird.feature.notification.testing.fake.FakeInAppOnlyNotification;
 import net.thunderbird.feature.notification.testing.fake.FakeNotificationManager;
+import net.thunderbird.legacy.core.StubLocalDeleteOperationDecider;
 import net.thunderbird.legacy.core.mailstore.folder.FakeOutboxFolderManager;
 import org.junit.After;
 import org.junit.Before;
@@ -131,6 +133,7 @@ public class MessagingControllerTest extends K9RobolectricTest {
         appContext = RuntimeEnvironment.getApplication();
 
         preferences = Preferences.getPreferences();
+        final LocalDeleteOperationDecider noOpLocalDeleteOperationDecider = new StubLocalDeleteOperationDecider();
         featureFlagProvider = key -> Disabled.INSTANCE;
 
         final NotificationManager notificationManager = new FakeNotificationManager(
@@ -151,7 +154,7 @@ public class MessagingControllerTest extends K9RobolectricTest {
             messageStoreManager,
             saveMessageDataCreator,
             specialLocalFoldersCreator,
-            new LocalDeleteOperationDecider(),
+            noOpLocalDeleteOperationDecider,
             Collections.<ControllerExtension>emptyList(),
             featureFlagProvider,
             syncLogger,
